refactor: simplify tombstone tests per review feedback
Move tests from source mod to creep mod, remove internal scratch
storage manipulation, and use only the public simulation API
(tick/peekRoom). Drop keeper lair tests that were outside the
scope of the tombstone fix.
