+++ src/mods/source/index.ts
@@ -7,5 +7,5 @@ export const manifest: Manifest = {
 		'xxscreeps/mods/resource',
 		'xxscreeps/mods/structure',
 	],
-	provides: [ 'backend', 'constants', 'game', 'processor', 'test' ],
+	provides: [ 'backend', 'constants', 'game', 'processor' ],
 };
